+++ packages/plugin-dot-net-core/__tests__/index.js
