+++ packages/plugin-dot-net/__tests__/index.js
