+++ packages/plugin-go/__tests__/index.js
