+++ packages/plugin-haskell/__tests__/index.js
