+++ packages/plugin-homebrew-manifest/__tests__/index.js
