+++ packages/plugin-java/__tests__/index.js
