+++ packages/plugin-javascript/__tests__/index.js
