+++ packages/plugin-rust/__tests__/index.js
