+++ packages/plugin-vim/__tests__/index.js
